Move starRating component to components folder, it might be used elsewhere.
