+++ src/components/log-dialog/log-route.tsx
@@ -10,7 +10,7 @@ import IconDelete from "../ui/icons/delete";
 import IconCollapse from "../ui/icons/collapse";
 import IconExpand from "../ui/icons/expand";
 import { difficultyToGrade } from "@/utils/grade-helpers";
-import StarRating from "./star-rating";
+import StarRating from "../star-rating";
 import {
   TLogRoute,
   tickAscentTypes,
